Fix stopwatch overflow bug
The subtraction needs to be cast to a signed integer with the same
number of bits, or the trick doesn't work.

+++ clock.cpp
@@ -83,7 +83,7 @@ namespace clock {
                 curTime = time_;
             }
 
-            long secondsRemaining = (long) (endTime_ - curTime);
+            int secondsRemaining = (int) (endTime_ - curTime);
 
             if (secondsRemaining > 0) {
                 return (unsigned) secondsRemaining;
